Stock-Web-API/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
jsonschema==2.6.0
requests==2.23.0
requests-oauthlib==1.3.0
python-dateutil==2.6.0
isodate==0.6.0